Add simple top-level query capability
Make simple queries translate directly to JQL at compile time. Nesting
should be simple enough to implement by replacing dots with slashes and
embedding last part in the squares. For instance `a.b.c = "d"` becomes
`/a/b[c = "d"]` and then multiple conditions can be gruped with and/or.